feat: fix filters without fixing aside block
