+++ src/files/products.json
@@ -140,7 +140,7 @@
         "stock": 14,
         "brand": "LAMBORGHINI",
         "category": "велосипед трехколесный",
-        "latinCategory": "Childbike",
+        "latinCategory": "tricycle",
         "thumbnail": "../assets/images/list/main/11.jpg",
         "images": ["../assets/images/list/mini/11.1.jpg", "../assets/images/list/mini/11.2.jpg"]
       },
@@ -153,7 +153,7 @@
         "stock": 18,
         "brand": "Safari proff",
         "category": "велосипед трехколесный",
-        "latinCategory": "Childbike",
+        "latinCategory": "tricycle",
         "thumbnail": "../assets/images/list/main/12.jpg",
         "images": ["../assets/images/list/mini/12.1.jpg", "../assets/images/list/mini/12.2.jpg"]
       },
@@ -166,7 +166,7 @@
         "stock": 27,
         "brand": "Tech Team",
         "category": "велосипед трехколесный",
-        "latinCategory": "Childbike",
+        "latinCategory": "tricycle",
         "thumbnail": "../assets/images/list/main/13.jpg",
         "images": ["../assets/images/list/mini/13.1.jpg", "../assets/images/list/mini/13.2.jpg"]
       },
