+++ src/modules/modal.ts
@@ -6,7 +6,6 @@ export function closeModal() {
   const element = document.querySelector('.modal');
   if (element) {
     element.classList.remove('slideInDown');
-    element.classList.add('slideInUp');
     element.remove();
   }
   document.body.style.overflow = '';
@@ -15,7 +14,6 @@ export function closeModal() {
 
 const createModal = () => {
   const scroll = calcScroll();
-
 
   modal.classList.add('modal', 'slideInDown');
   modal.setAttribute('id', 'exampleModal2');
